+++ js/main.js
@@ -1,30 +1,11 @@
-function getRandomPositiveInteger(a, b) {
-  if (a < 0 || b < 0) {
-    return NaN;
-  }
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-}
-
-let indexID = 0;
-
-const createId = function () {
-  indexID++;
-  return indexID;
-};
-
-function createPhoto() {
-  return {
-    id: createId(),
-    url: 'photos / { indexID }.jpg',
-    description: 'Описание фотографии',
-    likes: getRandomPositiveInteger(15, 200),
-    comments: getRandomPositiveInteger(15, 200),
-  };
-}
-
-const resultArray = Array.from({ length: 25 }, createPhoto);
+/*
+    Импорты других модулей
+    Вызовы общих функций
+    Настройка скриптов
+    ...
+*/
+import './util.js';
+import './data.js';
+import {resultArray} from './data.js';
 
 console.log(resultArray);
